+++ spec/features/alliances_spec.rb
@@ -18,7 +18,7 @@ describe 'List all active player alliances' do
 
     specify { expect(subject.alliance_ids.first).to eq(1_354_830_081) }
 
-    specify { expect(subject.current_etag).to eq('97f0c48679f2b200043cdbc3406291fc945bcd652ddc7fc11ccdc37a') }
+    specify { expect(subject.etag).to eq('97f0c48679f2b200043cdbc3406291fc945bcd652ddc7fc11ccdc37a') }
   end
 
   context 'when etag is set' do
@@ -30,6 +30,6 @@ describe 'List all active player alliances' do
 
     specify { expect(subject.not_modified?).to eq(true) }
 
-    specify { expect(subject.current_etag).to eq('97f0c48679f2b200043cdbc3406291fc945bcd652ddc7fc11ccdc37a') }
+    specify { expect(subject.etag).to eq('97f0c48679f2b200043cdbc3406291fc945bcd652ddc7fc11ccdc37a') }
   end
 end