+++ lib/eve_online/esi/universe_races.rb
@@ -2,7 +2,7 @@
 
 module EveOnline
   module ESI
-    class Races < Base
+    class UniverseRaces < Base
       API_ENDPOINT = 'https://esi.evetech.net/v1/universe/races/?datasource=%<datasource>s'
 
       def races