+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Add `EveOnline::ESI::Alliances#etag`
 * Add `etag` to `EveOnline::ESI::Models::Alliance#as_json`
 * Add `EveOnline::ESI::Alliance#etag`
+* Rename `EveOnline::ESI::Races` to `EveOnline::ESI::UniverseRaces`
 
 ## v0.20.0